MNT: Move fitting/ to core/fitting/
