+++ skxray/core/fitting/base/parameter_data.py
