+++ skxray/core/fitting/xrf_model.py
@@ -50,15 +50,16 @@ from collections import OrderedDict
 import logging
 
 import numpy as np
+
 from scipy.optimize import nnls
 import six
 
 from skxray.core.constants import XrfElement as Element
-from skxray.fitting.lineshapes import gaussian
-from skxray.fitting.models import (ComptonModel, ElasticModel,
-                                   _gen_class_docs)
-import skxray.fitting.base.parameter_data as sfb_pd
-from skxray.fitting.background import snip_method
+from skxray.core.fitting.lineshapes import gaussian
+from skxray.core.fitting.models import (ComptonModel, ElasticModel,
+                                        _gen_class_docs)
+from .base import parameter_data as sfb_pd
+from .background import snip_method
 from lmfit import Model
 
 logger = logging.getLogger(__name__)
